requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-postgres
kafka-python
pandas
psycopg2
joblib
apache-airflow
apache-airflow-providers-snowflake